requirements.txt
alembic==1.18.1
blinker==1.9.0
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
email-validator==2.1.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.3.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
mysqlclient==2.2.4
PyMySQL==1.1.2
python-dotenv==1.0.0
SQLAlchemy==2.0.45
tomli==2.4.0
typing_extensions==4.15.0
Werkzeug>=3.1
WTForms==3.1.1
Pillow>=9.0.0
qrcode[pil]>=7.0
pytz==2023.3